+++ mpqc/src/lib/util/group/memarmci.h
@@ -54,9 +54,9 @@ class ARMCIMemoryGrp: public RDMAMemoryGrp {
 
     void set_localsize(size_t);
 
-    void retrieve_data(void *, int node, int offset, int size, int lock);
-    void replace_data(void *, int node, int offset, int size, int unlock);
-    void sum_data(double *data, int node, int doffset, int dsize);
+    void retrieve_data(void *, int node, long offset, long size, int lock);
+    void replace_data(void *, int node, long offset, long size, int unlock);
+    void sum_data(double *data, int node, long doffset, long dsize);
 
     void sync();
     void deactivate();
